Fixes #55: add support for unstable merge status
Sometimes you have broken actions that are not because of you. My
example is Vercel :angry: failing my PR just because I don't have
a user in some repo.

Anyways, here's a fix - added a commandline switch with the merge
mode. The default is "Clean" which is what you had before, so people
who already use the tool won't feel the difference.

Signed-off-by: Shay Nehmad <dude500@gmail.com>