Mark PRs temporarily draft while submitting (#111)
When running `stack-pr submit`, patches in a stack need to have their base branches temporarily changed to avoid accidental closures. This has the unfortunate side effect that in this transient state, the PRs seem like multi-commit PRs pointing to main, resulting in GH notifying reviewers for every commit on the branch (as opposed to just the one that corresponds to the stacked PR). This patch fixes this by marking PRs temporarily draft while we manipulate their bases.